+++ src/main/java/scanner/http/HttpClient.java
@@ -7,6 +7,11 @@ import okhttp3.Response;
 
 import java.util.concurrent.TimeUnit;
 
+/**
+ * Simple http client.
+ *
+ * @author inkarnadin
+ */
 public class HttpClient {
 
     private final static OkHttpClient client = new OkHttpClient()
@@ -15,6 +20,12 @@ public class HttpClient {
             .callTimeout(2, TimeUnit.SECONDS)
             .build();
 
+    /**
+     * Execute usual request.
+     *
+     * @param path url
+     * @return response
+     */
     @SneakyThrows
     public static Response execute(String path) {
         Request request = new Request.Builder()
@@ -23,6 +34,13 @@ public class HttpClient {
         return client.newCall(request).execute();
     }
 
+    /**
+     * Execute basic auth request.
+     *
+     * @param path url
+     * @param auth credentials as base64 string
+     * @return response
+     */
     @SneakyThrows
     public static Response executeBasicAuth(String path, String auth) {
         Request request = new Request.Builder()