+++ src/test/java/scanner/CameraScannerTest.java
@@ -9,8 +9,8 @@ public class CameraScannerTest {
     public void testPrepareSinglePortScanning() {
         String range = "188.202.63.40-188.202.63.47";
 
-        CameraRangeResolver.prepareSinglePortScanning(range);
-        Assert.assertEquals(8, CameraRangeResolver.count());
+        CameraRangeManager.prepareSinglePortScanning(range);
+        Assert.assertEquals(8, CameraRangeManager.count());
     }
 
 }