+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Added flag `-w` waiting timeout host response (sec, default 200 ms).
+- Added support for large ip ranges.
 
 ### Changed
 - Updated special RTSP path as 11 instead Streaming/Channels/101.
@@ -13,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Changed the mechanism of counting all IP addresses that will be scanned.
 - Changed unit for timeout flags - `-w` and `-bw` for milliseconds instead second.
 - Changed default active threads value for 20 instead 10.
+- Optimized ip range working.
 
 ### Fixed
 - Fixed some log naming typos.